OTA: Initial commit
This repository is meant to serve as an official OTA server for crDroid updates in the future. New releases will be automatically uploaded here and users will be able to update from the updater app! However this needs ROM-side updates. So for now, this just points to the most recent release.